[portmgrd] regression: prevent runtime exception (crash) in configuring portchannel at boot (#3432)
* portmgrd: prevent runtime failure in setting MTU on portchannel member

Prevent setting the PORT MTU on PortChannel members as it will likely fail
and cause portmgrd to exit (The PortChannel itself is where the MTU gets set,
not the PORT).  The current code is setting a default value for an MTU (9100)
even when its a PortChannel member, so this patch prevents that default value
from being set.  Also if a user were to incorrectly specify an MTU on a Port
that is a member of the port channel via `config_db.json` this too would bring
down portmgrd, so catch that and just emit a warning instead.